Adds standalone tests
