+++ snap_plugin/v1/tests/test_collector.py
@@ -17,7 +17,6 @@
 
 import json
 import sys
-import threading
 import time
 from builtins import int as bigint
 
@@ -30,76 +29,7 @@ from snap_plugin.v1.metrics_arg import MetricsArg
 from snap_plugin.v1.plugin_pb2 import CollectorStub, Empty
 from snap_plugin.v1.tests import ThreadPrinter
 
-
-class MockCollector(snap.Collector, threading.Thread):
-    """Mock collector plugin """
-
-    def __init__(self, name, ver):
-        super(MockCollector, self).__init__(name, ver)
-        self._flags.add('require-config', snap.plugin.FlagType.toggle, '')
-        threading.Thread.__init__(self, group=None, target=None, name=None)
-        self._stopper = threading.Event()
-
-    def collect(self, metrics):
-        for metric in metrics:
-            metric.timestamp = time.time()
-            metric.version = 2
-            if "bytes" in metric.config and metric.config["bytes"] is True:
-                metric.data = b'qwerty'
-            elif "string" in metric.config and metric.config["string"] is True:
-                metric.data = "qwerty"
-            elif "int32" in metric.config and metric.config["int32"] is True:
-                metric.data = 99
-            elif "int64" in metric.config and metric.config["int64"] is True:
-                metric.data = bigint(99)
-            elif "bool" in metric.config and metric.config["bool"] is True:
-                metric.data = True
-            else:
-                metric.data = 99.9
-        return metrics
-
-    def update_catalog(self, config):
-        now = time.time()
-        metrics = [
-            snap.Metric(
-                namespace=[
-                    snap.NamespaceElement(value="acme"),
-                    snap.NamespaceElement(value="sk8"),
-                    snap.NamespaceElement(value="matix")
-                ],
-                unit="some unit",
-                description="some description",
-                timestamp=now,
-            )
-        ]
-        return metrics
-
-    def get_config_policy(self):
-        policy = [
-            ("acme", "sk8", "matix"),
-            [
-                (
-                    "password",
-                    snap.StringRule(default="grace", required=True),
-                ),
-                (
-                    "user",
-                    snap.StringRule(default="kristy", required=True),
-                ),
-            ]
-        ]
-
-        if self._args.require_config:
-            policy[1].append(("database", snap.StringRule(required=True)))
-
-        return snap.ConfigPolicy(policy)
-
-    def run(self):
-        self.start_plugin()
-
-    def stop(self):
-        self._stopper.set()
-        self.stop_plugin()
+from .mock_plugins import MockCollector
 
 
 @pytest.fixture(scope="module")
